Update OrderParameter2D.m
Line 349 of the OrderParameter2D panel (in the move_and_rotate function) was taking the difference of original fiber positions, rather than the moved and rotated points. This led to a S_rand that converged to the same value as S_im at maximum frame size.

diff(xy, 1, 2);
has been changed to 
diff(p_rand, 1, 2)

and this leads to normal behavior.